Rename DefaultWorkRunners to MobiusPlugins
